eye_clinic_app3/requirements.txt
mysql-connector-python==8.0.28
python-dotenv==0.19.2